add alert message add to cart and delete from cart

+++ src/components/buttonDelete.jsx
@@ -5,6 +5,7 @@ import { checkResponseStatus } from "../functions/checkResponseStatus";
 import { deleteFromCart } from "../redux/CartSlice";
 import { useNavigate } from "react-router-dom";
 import { useDispatch } from "react-redux";
+import Swal from "sweetalert2";
 
 const ButtonDelete = ({element}) => {
     const [load,setLoading] = useState(false)
@@ -12,19 +13,35 @@ const ButtonDelete = ({element}) => {
     const dispatch = useDispatch();
     const deleCart = (ele) => {
         setLoading(true)
-        axios
-        .delete(`https://ecommerce-api-hlp7.onrender.com/api/cart/${ele._id}`, {
-            headers: {
-            Authorization: `Bearer ${localStorage.token}`,
-            },
-        })
-        .then(() => {
-            dispatch(deleteFromCart(ele))
-            setLoading(false)
-        })
-        .catch((e) => {
-            checkResponseStatus(e, myUrl);
+        Swal.fire({
+            title: "Are you sure?",
+            text: "You won't be able to revert this!",
+            icon: "warning",
+            showCancelButton: true,
+            confirmButtonColor: "#d33",
+            cancelButtonColor: "#3085d6",
+            confirmButtonText: "Yes, delete it!"
+        }).then((result) => {
+            if (result.isConfirmed) {
+                axios.delete(`https://ecommerce-api-hlp7.onrender.com/api/cart/${ele._id}`, {
+                    headers: {
+                    Authorization: `Bearer ${localStorage.token}`,
+                    },
+                })
+                .then(() => {
+                    dispatch(deleteFromCart(ele))
+                    Swal.fire({
+                        title: "Deleted!",
+                        text: "Your Book has been deleted.",
+                        icon: "success"
+                    });
+                })
+                .catch((e) => {
+                    checkResponseStatus(e, myUrl);
+                });
+            }
         });
+        setLoading(false)
     };
     return (
         <button disabled={load}
@@ -33,7 +50,7 @@ const ButtonDelete = ({element}) => {
             {
                 load ?
                     <div className="flex justify-center">
-                        <span className="inline-block border-2 border-white rounded-full w-5 h-5 border-l-gray-500 animate-spin"></span>
+                        <span className="inline-block w-5 h-5 border-2 border-white rounded-full border-l-gray-500 animate-spin"></span>
                     </div>
                 :"Delete"
                 }
